Refactored FileIO code to another class that streamlines all such similar code.
