+++ game_stats.py
@@ -3,7 +3,7 @@ class GameStats:
 
     def __init__(self, game_instance):
         """Initialize statistics."""
-        self.high_score = self.input_high_score()
+        self.high_score = game_instance.highscore_io.input_high_score()
 
         # Setup fully dynamic stats
         self.reset_stats(game_instance)
@@ -19,13 +19,3 @@ class GameStats:
         self.game_active = False
         self.play_button_clicked = False
         self.difficulty_button_clicked = False
-
-    def input_high_score(self) -> int:
-        """Input high score from output/high_score.txt if it exists"""
-        input_dir = "output/high_score.txt"
-
-        try:  # File exists
-            with open(input_dir, 'r') as f:
-                return int(float(f.read()))
-        except FileNotFoundError:  # Default value
-            return 0
